Reorganize name resolution
