+++ crates/ra_hir/src/code_model_api.rs
@@ -8,8 +8,7 @@ use crate::{
     Name, ScopesWithSourceMap, Ty, HirFileId,
     HirDatabase, PersistentHirDatabase,
     type_ref::TypeRef,
-    nameres::{ModuleScope, Namespace, lower::ImportId},
-    nameres::crate_def_map::ModuleId,
+    nameres::{ModuleScope, Namespace, ImportId, ModuleId},
     expr::{Body, BodySourceMap},
     ty::InferenceResult,
     adt::{EnumVariantId, StructFieldId, VariantDef},
